Refactor interactive widgets, fix URL clicker bug, add stroke on colors
